Align browse and search orchestration
