+++ apps/template/components/collections/collection-page.tsx
@@ -1,15 +1,9 @@
-import { SlidersHorizontalIcon } from "lucide-react";
 import Link from "next/link";
 import { Suspense } from "react";
 
 import { CollectionViewedTracker } from "@/components/analytics/trackers";
-import { FilterSidebarSheet } from "@/components/collections/filter-sidebar-sheet";
-import { CollectionFilters } from "@/components/collections/filters";
 import { CollectionResultsGrid } from "@/components/collections/results-grid";
-import { CollectionsSortSelect } from "@/components/collections/sort-select";
-import { SortSelectFallback } from "@/components/collections/sort-select-fallback";
-import { CollectionToolbar } from "@/components/collections/toolbar";
-import { ProductsGridSkeleton } from "@/components/product/products-grid";
+import { BrowseFallback, BrowseToolbar } from "@/components/collections/toolbar";
 import { BreadcrumbSchema } from "@/components/schema/breadcrumb-schema";
 import { CollectionSchema } from "@/components/schema/collection-schema";
 import { Container } from "@/components/ui/container";
@@ -18,10 +12,7 @@ import { Sections } from "@/components/ui/sections";
 import type { CollectionResultsData, CollectionSearchState } from "@/lib/collections/server";
 import type { Collection } from "@/lib/types";
 
-import {
-  CollectionActiveFilterCountBadge,
-  CollectionBrowseProvider,
-} from "./collection-browse-provider";
+import { CollectionBrowseProvider } from "./collection-browse-provider";
 import { FilterPendingScope } from "./filter-pending-context";
 
 export function CollectionDetailPage({
@@ -37,8 +28,6 @@ export function CollectionDetailPage({
   searchStatePromise: Promise<CollectionSearchState>;
   sortExclude?: string[];
 }) {
-  const filtersLabel = "Filters";
-  const sortByLabel = "Sort";
   return (
     <>
       {collection.id ? (
@@ -49,37 +38,13 @@ export function CollectionDetailPage({
           <Sections className="gap-5">
             <CollectionHeader collection={collection} handle={handle} homeLabel="Home" />
 
-            <Suspense
-              fallback={
-                <CollectionBrowseFallback filtersLabel={filtersLabel} sortByLabel={sortByLabel} />
-              }
-            >
+            <Suspense fallback={<BrowseFallback />}>
               <CollectionBrowseProvider handle={handle} searchStatePromise={searchStatePromise}>
-                <CollectionToolbar
-                  filterSheet={
-                    <FilterSidebarSheet
-                      label={filtersLabel}
-                      trigger={
-                        <button
-                          type="button"
-                          className="flex items-center gap-2 text-sm font-medium"
-                        >
-                          <SlidersHorizontalIcon className="size-4" />
-                          <span>{filtersLabel}</span>
-                          <CollectionActiveFilterCountBadge />
-                        </button>
-                      }
-                    >
-                      <FilterPendingScope>
-                        <CollectionFilters
-                          facetsPromise={collectionResultsDataPromise.then(
-                            (data) => data.transformedFilters,
-                          )}
-                        />
-                      </FilterPendingScope>
-                    </FilterSidebarSheet>
-                  }
-                  sortSelect={<CollectionsSortSelect exclude={sortExclude} />}
+                <BrowseToolbar
+                  facetsPromise={collectionResultsDataPromise.then(
+                    (data) => data.transformedFilters,
+                  )}
+                  sortExclude={sortExclude}
                 />
 
                 <FilterPendingScope>
@@ -92,29 +57,6 @@ export function CollectionDetailPage({
           </Sections>
         </Container>
       </Page>
-    </>
-  );
-}
-
-function CollectionBrowseFallback({
-  filtersLabel,
-  sortByLabel,
-}: {
-  filtersLabel: string;
-  sortByLabel: string;
-}) {
-  return (
-    <>
-      <CollectionToolbar
-        filterSheet={
-          <button type="button" className="flex items-center gap-2 text-sm font-medium">
-            <SlidersHorizontalIcon className="size-4" />
-            <span>{filtersLabel}</span>
-          </button>
-        }
-        sortSelect={<SortSelectFallback label={sortByLabel} />}
-      />
-      <ProductsGridSkeleton count={40} className="sm:grid-cols-3 lg:grid-cols-4 xl:grid-cols-5" />
     </>
   );
 }
